Add failing tests for constructor ordering bug
It happens when there are two constructors and the first has more than
one parameter with none optional and second has the same parameters as
the first plus an additional with all parameters optional. When the
available configuration fields consists of a single item that matches
one of the shared parameters, then the illegal first constructor is
selected, when the legal second constructor should be selected.