Add failing tests for constructor ordering bug
It happens when there are two constructors with the same number of
parameters where one constructor has all default parameters and the
other has one non-default parameter. The non-default parameter has a
match in the config, but none of the default parameters do. It is
currently declaring both constructors equal, but the one that has the
specified parameter should be selected.